Téléchargement d'image

+++ controllers/books.js
@@ -1,26 +1,61 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
 exports.createBook = (req, res, next) => {
-  delete req.body._id;
-  const book = new Book({
-    ...req.body
-  });
-  book.save()
-    .then(() => res.status(201).json({ message: 'Livre ajouté !'}))
-    .catch(error => res.status(400).json({ error }));
-}
+   const bookObject = JSON.parse(req.body.thing);
+   delete bookObject._id;
+   delete bookObject._userId;
+   const book = new Book({
+       ...bookObject,
+       userId: req.auth.userId,
+       imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+   });
+ 
+   book.save()
+   .then(() => { res.status(201).json({message: 'Livre enregistré !'})})
+   .catch(error => { res.status(400).json( { error })})
+};
 
 exports.modifyBook = (req, res, next) => {
-  Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Livre modifié !'}))
-    .catch(error => res.status(400).json({ error }));
-}
+   const bookObject = req.file ? {
+       ...JSON.parse(req.body.thing),
+       imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+   } : { ...req.body };
+ 
+   delete bookObject._userId;
+   Book.findOne({_id: req.params.id})
+       .then((book) => {
+           if (book.userId != req.auth.userId) {
+               res.status(401).json({ message : 'Non autorisé' });
+           } else {
+               Book.updateOne({ _id: req.params.id}, { ...bookObject, _id: req.params.id})
+               .then(() => res.status(200).json({message : 'Livre modifié!'}))
+               .catch(error => res.status(401).json({ error }));
+           }
+       })
+       .catch((error) => {
+           res.status(400).json({ error });
+       });
+};
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Livre supprimé !'}))
-    .catch(error => res.status(400).json({ error }));
-}
+   Book.findOne({ _id: req.params.id})
+       .then(book => {
+           if (book.userId != req.auth.userId) {
+               res.status(401).json({message: 'Non autorisé'});
+           } else {
+               const filename = book.imageUrl.split('/images/')[1];
+               fs.unlink(`images/${filename}`, () => {
+                   Book.deleteOne({_id: req.params.id})
+                       .then(() => { res.status(200).json({message: 'Livre supprimé !'})})
+                       .catch(error => res.status(401).json({ error }));
+               });
+           }
+       })
+       .catch( error => {
+           res.status(500).json({ error });
+       });
+};
 
 exports.getOneBook =  (req, res, next) => {
   Book.findOne({ _id: req.params.id })
